requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
django-filter==2.4.0
flake8==3.9.2
idna==3.2
importlib-metadata==4.6.3
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1 
python-dotenv
djoser
docxtpl